PYBIND: It seems to compile now, and run.
However, I don't understand why the algorithm::
is used for BFS.  NV can perhaps fix this.

CHANGE: Had to include -fPIC for library compilation
in order for the code to be linked with pybind.

TODO: There is also a warning on add_library, but
I can look into fixing this later.